% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{download_organization_repos}
\alias{download_organization_repos}
\title{Download archives of GitHub repositories within an organization}
\usage{
download_organization_repos(
  organization = NULL,
  extra_repos = NULL,
  dest_folder = getwd()
)
}
\arguments{
\item{organization}{Organization name}

\item{extra_repos}{Vector of extra repository names}

\item{dest_folder}{Where to save the folders with the archives.}
}
\description{
Download archives of GitHub repositories within an organization
}
\details{
You will need a \href{https://usethis.r-lib.org/articles/git-credentials.html}{GitHub Personal Access Token}.

The reason why you might need \code{extra_repos} is if you want to download archives
of private repositories, while using a GitHub Personal Access Token with no scope.

\href{https://docs.github.com/en/repositories/archiving-a-github-repository/backing-up-a-repository}{GitHub documentation on archives}.
}
\examples{
\dontrun{
download_organization_repos("maelle-test")
}
}
