% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{download_organization_repos}
\alias{download_organization_repos}
\title{Download archives of GitHub repositories within an organization}
\usage{
download_organization_repos(
  organizations = NULL,
  extra_repos = NULL,
  keep = character(0),
  dest_folder = getwd()
)
}
\arguments{
\item{organizations}{Organization name(s) (vector)}

\item{extra_repos}{Named vector of extra repository names where names are organization names.}

\item{keep}{a character vector of what files to keep. If this vector is length zero, all repositories are kept.}

\item{dest_folder}{Where to save the folders with the archives.}
}
\description{
Download archives of GitHub repositories within an organization
}
\details{
You will need a \href{https://usethis.r-lib.org/articles/git-credentials.html}{GitHub Personal Access Token}.

The reason why you might need \code{extra_repos} is if you want to download archives
of private repositories, while using a GitHub Personal Access Token with no scope.

\href{https://docs.github.com/en/repositories/archiving-a-github-repository/backing-up-a-repository}{GitHub documentation on archives}.
}
\examples{
\dontrun{
download_organization_repos(c("maelle-test", "maelle-test"))
download_organization_repos("maelle-test", keep = "testy2") # only keep the testy2 repo
}
}
